fix(mobile): satisfy deep-link file size gate
