+++ mobile/test/features/channels/message_content/link_normalizer_test.dart
@@ -1,0 +1,36 @@
+import 'package:buzz/features/channels/message_content/link_normalizer.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  const url = 'buzz://message?channel=channel-1&id=message-1';
+
+  test('normalizes supported bare and autolinked Buzz URLs', () {
+    expect(
+      normalizeBareLinks('See $url and <$url>'),
+      'See [$url]($url) and [$url]($url)',
+    );
+  });
+
+  test('keeps punctuation and open Markdown delimiters outside links', () {
+    expect(
+      normalizeBareLinks('**open $url**. and **_${url}_**!'),
+      '**open [$url]($url)**. and **_[$url]($url)_**!',
+    );
+  });
+
+  test('preserves URL suffix characters without a matching opener', () {
+    expect(
+      normalizeBareLinks(
+        'See $url'
+        '_ and $url~~',
+      ),
+      'See [$url'
+      '_]($url'
+      '_) and [$url~~]($url~~)',
+    );
+  });
+
+  test('leaves links inside backticks untouched', () {
+    expect(normalizeBareLinks('`$url` then $url'), '`$url` then [$url]($url)');
+  });
+}
